backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.1
pyzmq==27.1.0
GitPython==3.1.45
SQLAlchemy==2.0.36
psycopg[binary]==3.2.3
pyjwt==2.10.1
pwdlib[argon2]==0.3.0